.TIMES in class template mocks with self expectation
In the odd case that a class template is a mock, and where
there are functions in the class template that sets
expectations on mocked functions and uses .TIMES(), the
call of the .times<>() member function needs to take the
multiplicity as parameters and not using explicit template
arguments. The error messages given by clang++ and g++
were very unsupporting for pinpointing the actual problem.