More review comments
I started to add the subheadings, but I backed it out because the local and global install share the "use in cmake" part, which breaks the separateness of the sections. I haven't used the `add_subdirectory()` method myself, so I just added it in passing at the bottom.

I propose you accept my pull request and then make any further elaborations you feel are necessary. :-)