Allow IMPLEMENT_MOCK() for protected base methods
The function signature type passed into the `TROMPELOEIL_MAKE_MOCK_` macro was repeated within
nested structures, which did not have access to protected methods in the actual base class being
mocked. This solution stores the `sig` parameter to an alias, and that alias is instead used in the
nested struct. This allows `INTERFACE_MOCK` to be used for protected base class members.