Add tests for TROMPELOEIL_*_STDMETHOD_MOCK
As these macros are expected to work with COM interfaces, we mock the
IDispatch interface as a test, this brings windows headers into the test
which unfourtunetly has to come before the trompeloeil.hpp header to
work. It also defines "interface" as a macro, but we don't need it so
we can undefine it.

We also add the same test as the other  MAKE_*_MOCKn macros to test the
15 variants.