Add MSVC configuration with /Zc:__cppstd set to c++20
Demonstrate bug #343 - since MSVC offers individual options to control
/std:c++20, /Zc:__cplusplus, and /Zc:preprocessor, it's possible to have
a combination where the compiler is in C++20 mode (and says so), but
the traditional preprocessor (which not support __VA_OPT__) is in use.

This configuration is commonly trigged by Qt6, which requires c++17,
suggests c++20, sets /Zc:__cpluspls (so it can #ifdef based on compiler),
but doesn't care about the preprocessor. MSVC still doesn't enable the
new /Zc:preprocessor by default, even with /std:c++20 or /permissive-.